add new base class for Newgame, that doesn't depend on agents
